+++ src/models/google-tag-manager-config.ts
@@ -1,0 +1,12 @@
+export interface GoogleTagManagerConfig {
+  containerId: string;
+  customEventsTracking: {
+    custom: string[];
+    preset: {
+      coreEvents: boolean;
+      UIEvents: boolean;
+      playlistEvents: boolean;
+      castEvents: boolean;
+    };
+  };
+}
